backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
ascii_colors==0.5.2
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
lightrag-hku==1.1.4
MarkupSafe==3.0.2
nano-vectordb==0.0.4.3
neo4j==5.27.0
networkx==3.4.2
numpy==2.2.2
openai==1.61.0
pipmaster==0.4.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3

accelerate
aiofiles
aiohttp
configparser

# File manipulation libraries
docling
graspologic

# database packages
networkx

# Basic modules
numpy
pipmaster
pydantic
python-dotenv

setuptools
tenacity


# LLM packages
tiktoken
tqdm
xxhash